Only copy failsafe artifacts if they exist.

+++ circle.sh
@@ -38,8 +38,10 @@ case "$1" in
 
     collect_test_reports)
         mkdir -p $CIRCLE_TEST_REPORTS/surefire
-        mkdir -p $CIRCLE_TEST_REPORTS/failsafe
         cp target/surefire-reports/TEST-*.xml $CIRCLE_TEST_REPORTS/surefire
-        cp target/failsafe-reports/TEST-*.xml $CIRCLE_TEST_REPORTS/failsafe
+        if [ -e target/failsafe-reports ] ; then
+            mkdir -p $CIRCLE_TEST_REPORTS/failsafe
+            cp target/failsafe-reports/TEST-*.xml $CIRCLE_TEST_REPORTS/failsafe
+        fi
         ;;
 esac
